started tests, added a lib to ker-server
